backend/pyproject.toml
[tool.poetry]
name = "accelerator"
version = "0.1.0"
description = "Django backend"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
django-htmx = "^1.19.0"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.2"
django-render-block = "^0.10"
pillow = "^10.4.0"
gunicorn = "^21.2"
django-cors-headers = "^4.7.0"
drf-spectacular = "^0.28.0"
djangorestframework-simplejwt = "^5.5.0"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
factory-boy = "^3.3.3"
faker = "^37.6.0"
aiohttp = "^3.12.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"